pyproject.toml
[project]
name = "some_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "dotenv==0.9.9",
    "fastapi==0.118.2",
    "pyyaml==6.0.3",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "uvicorn==0.37.0",
    "pydantic==2.12.0",
    #    "flash-attn @ file://${PROJECT_ROOT}/wheels/flash_attn-2.8.3+cu12torch2.8cxx11abiTRUE-cp312-cp312-linux_x86_64.whl",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[python]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://mirror.sjtu.edu.cn/pytorch-wheels/cu126"
#url = "https://download.pytorch.org/whl/cu126"
explicit = true
